Update solution for sigmoid-numpy

+++ sigmoid-numpy/sigmoid-numpy.py
@@ -4,5 +4,5 @@ def sigmoid(x):
     """
     Vectorized sigmoid function.
     """
-    x = np.array(x)
+    x = np.array(x, dtype=float)
     return 1 / (1 + np.exp(-x))
